baseline/requirements-baseline.txt
openkiwi==2.1.0
scikit-learn==1.0.1
jsonlines==2.0.0
